backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
aiofiles>=23.2.1
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.23.0
openai>=0.27.0
Flask>=3.0.2
Werkzeug>=3.0.1
Jinja2>=3.1.2
click>=8.1.6
itsdangerous>=2.1.2
MarkupSafe>=2.1.5
blinker>=1.7.0
packaging>=24.0
typing_extensions>=4.10.0
pandas>=2.0.0
